requirements.txt
django_phonenumber_field==4.0.0
dj_database_url==0.5.0
channels==3.0.2
djangorestframework_jwt==1.11.0
django_filter==2.4.0
drf_jwt_2fa==0.3.0
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.4.0
django-phonenumber-field==4.0.0
django-querysetsequence==0.13
django-sequences==2.6
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
drf-jwt-2fa==0.3.0
gunicorn==20.0.4
html5lib==1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
msgpack==1.0.2
#mysqlclient==2.0.1
npyscreen==4.10.5
ordered-set==4.0.2
packaging==20.9
pexpect==4.8.0
phonenumberslite==8.12.7
Pillow==8.1.0
ply==3.11
progress==1.5
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
#pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyxdg==0.26
#PyYAML==5.4.1
reportlab==3.5.65
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
s3transfer==0.3.3
service-identity==18.1.0
setproctitle==1.2.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
txaio==20.4.1
udiskie==2.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
yapf==0.30.0
zope.interface==5.2.0
PyGObject==3.38.0
pycairo==1.20.0
Twisted==20.3.0
wrapt==1.12.1
